add importing from old dir to restricted imports
