+++ .eslintrc.json
@@ -38,7 +38,7 @@
     "no-only-tests/no-only-tests": "error",
     "no-restricted-imports": [
       "error",
-      { "paths": [".", "..", "../..", "../../.."] }
+      { "paths": [".", "..", "../..", "../../.."], "patterns": ["**/old/**"] }
     ],
     "simple-import-sort/imports": "error",
     "simple-import-sort/exports": "error",
