+++ packages/frontend/src/scripts/offer/filteringForm.ts
@@ -1,3 +1,4 @@
+// eslint-disable-next-line no-restricted-imports
 import {
   AssetIdSelectName,
   DisabledOptionValue,
