+++ packages/frontend/src/scripts/transaction/finalizeExitForm.ts
@@ -1,6 +1,7 @@
 import { encodeFinalizeExitRequest } from '@explorer/shared'
 import { EthereumAddress, Hash256, StarkKey } from '@explorer/types'
 
+// eslint-disable-next-line no-restricted-imports
 import { FormClass } from '../../view/old/forced-transactions/finalize-form'
 import { getAttribute } from '../offer/getAttribute'
