+++ packages/frontend/src/scripts/transaction/state.ts
@@ -1,5 +1,6 @@
 import { AssetId } from '@explorer/types'
 
+// eslint-disable-next-line no-restricted-imports
 import { TransactionFormProps } from '../../view/old/transaction-form'
 import { FormAction, FormState } from './types'
 import {
