+++ packages/frontend/src/scripts/transaction/transactionForm.ts
@@ -1,7 +1,10 @@
 import { AssetId } from '@explorer/types'
 
+// eslint-disable-next-line no-restricted-imports
 import { getAssetImageUrl } from '../../view/old/common/icons/getAssetImageUrl'
+// eslint-disable-next-line no-restricted-imports
 import { FormId } from '../../view/old/transaction-form/ids'
+// eslint-disable-next-line no-restricted-imports
 import { TransactionFormProps } from '../../view/old/transaction-form/TransactionFormProps'
 import { getFormElements } from './getFormElements'
 import { getInitialState, nextFormState } from './state'
