+++ packages/frontend/src/view/components/user/OffersTable.tsx
@@ -1,7 +1,7 @@
 import { Timestamp } from '@explorer/types'
 import React from 'react'
 
-import { NewTable } from '../common/table/NewTable'
+import { Table } from '../common/table/Table'
 import { Status, StatusType } from './Status'
 
 export interface OffersTableProps {
@@ -20,7 +20,7 @@ export interface OfferEntry {
 
 export function OffersTable({ offers }: OffersTableProps) {
   return (
-    <NewTable
+    <Table
       pageSize={6}
       id="test"
       title="Offers"
