+++ packages/frontend/src/view/index.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-restricted-imports */
 export * from './old/forced-transactions'
 export * from './old/home'
 export * from './old/not-found'
@@ -5,5 +6,6 @@ export * from './old/offers'
 export * from './old/positions'
 export * from './old/state-updates'
 export * from './old/transaction-form'
+/* eslint-enable no-restricted-imports */
 export * from './pages/HomePage'
 export * from './pages/UserPage'
