+++ packages/frontend/src/view/old/common/table/index.ts
@@ -1,4 +1,3 @@
-export * from '../../../components/common/table/NewTable'
 export * from './ClientPaginatedTable'
 export * from './Table'
 export * from './types'
